web/requirements.txt

pandas
sqlalchemy
python-dotenv
cx_Oracle


flask




groq